[modbus] Added options for reducing peak bus usage
While the modbus library is thread-save, it does not keep track of the
load of the bus and individual units. Therefore we introduce the two variables
force_delay and force_delay_per_unit, which enforce a time between two
consecutive bus interactions for the hole bus or just per unit respectively.