modbus: switch configuration to functionCode
The usage of accessType and HoldingRegister to select the operation
mode of modbus nodes has been clunky and unwieldy from the start.
In this patch we finally switch to directly entering the function
code in the config. This opens the possibility of easily adding
additional functioncodes in the future.

In order to keep compatibility with old configs, the accessType and
HoldingRegister variables can still be used in order to infer the
corresponding functionCode. However this is set to depricated and
will be removed in the future.